Split HearingEvent for Response and Result
The Result of the api now can place the data into its own HearingEvent model.
Same for the Response.

This seperation of concerns decouples the Result and the Response when storing the list of HearingEvent
