+++ laa_court_data_api_app/models/hearing_events/hearing_event/response/hearing_event.py
