+++ laa_court_data_api_app/models/hearing_events/hearing_events_response.py
@@ -3,7 +3,7 @@ from uuid import UUID
 
 from pydantic import BaseModel
 
-from laa_court_data_api_app.models.hearing_events.hearing_event import HearingEvent
+from laa_court_data_api_app.models.hearing_events.hearing_event.response.hearing_event import HearingEvent
 
 
 class HearingEventsResponse(BaseModel):
