+++ laa_court_data_api_app/models/hearing_events/hearing_events_result.py
@@ -3,8 +3,7 @@ from uuid import UUID
 
 from pydantic import BaseModel
 
-from laa_court_data_api_app.models.hearing_events.hearing_event import HearingEvent
-
+from laa_court_data_api_app.models.hearing_events.hearing_event.result.hearing_event import HearingEvent
 
 class HearingEventsResult(BaseModel):
     hearing_id: Optional[UUID] = None
